Add a model with independent dose weights
In this model, the time-lagged weights used to calculate the effective
dose are independent parameters drawn from a Dirichlet distribution.
This is achieved by adding a new model parameter "dose_weights", and
modifying the "compute_effective_dose()" function.

The model fits to the individual haemoglobin and reticulocyte data can
be plotted with the "plot_test_free_weights.R" script.

Note that the plotting script does not plot the individual fits in the
same order as per the existing analyses. I am unsure why, but perhaps it
is related to the missing "ID_map" variable? See the TODO notes in the
plotting script for details.