Only accept the initial haemocue measurement
In order to generate predictions for an input drug regimen, we only need
the individual's initial haemocue measurement.

Previously, the model required a measurement for each day of the input
drug regimen. However, this prevented us from generating predictions for
the entire study duration when there were missing data. The simplest
option was to pass only the initial measurement, making Y_true_haemocue
a scalar value rather than a vector.